move sections to app
