+++ apps/main/src/app/Sections/Stats.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Section from "../Section";
+import { Section } from "@repo/ui";
 
 const background = <div className="w-full h-full bg-[#BEACD0]" />;
