+++ packages/ui/package.json
@@ -9,7 +9,7 @@
   ],
   "exports": {
     "./styles.css": "./dist/index.css",
-    "./sections": "./src/Sections/index.ts",
+    ".": "./src/index.ts",
     "./components": "./src/Components/index.ts"
   },
   "license": "MIT",
