--- packages/ui/src/Sections/index.ts
@@ -1,2 +1,0 @@
-export { default as Section } from "./Section";
-export { Stats } from "./Stats";
