+++ packages/ui/src/index.ts
@@ -1,0 +1,2 @@
+export { default as Section } from "./Section";
+export type { ForegroundItems, Coordinate } from "./Section";
